requirements.txt
Django==4.2.0
djangorestframework==3.14.0
Pillow>=10.3.0
python-decouple==3.8
